+++ packages/account-core/src/constants.ts
@@ -13,11 +13,9 @@ import {
  * Chain IDs for the supported chains, only add chains here that are supported by the app on the basic level.
  * For mechanics like bridging, this is not the place to add them.
  */
-export type ChainId =
-  | typeof sophon.id
-  | typeof sophonTestnet.id
-  | typeof sophonOS.id
-  | typeof sophonOSTestnet.id;
+export type LegacyChainId = typeof sophon.id | typeof sophonTestnet.id;
+export type OSChainId = typeof sophonOS.id | typeof sophonOSTestnet.id;
+export type ChainId = LegacyChainId | OSChainId;
 
 export const SophonChains: Record<ChainId, Chain> = {
   [sophon.id]: sophon,
@@ -70,8 +68,8 @@ export const SophonChainCapabilities: Record<ChainId, ChainCapability> = {
     transactions: ChainCapabilityValue.DISABLED,
   },
   [sophonOSTestnet.id]: {
-    signature: ChainCapabilityValue.DISABLED,
-    deployContract: ChainCapabilityValue.OFF_CHAIN,
+    signature: ChainCapabilityValue.ENABLED,
+    deployContract: ChainCapabilityValue.ENABLED,
     sns: ChainCapabilityValue.DISABLED,
     transactions: ChainCapabilityValue.DISABLED,
   },
@@ -92,7 +90,7 @@ export const AccountAuthAPIURL: Record<ChainId, string> = {
   [sophon.id]: 'https://auth.sophonauth.com',
   [sophonTestnet.id]: 'https://auth.staging.sophonauth.com',
   [sophonOS.id]: 'https://auth.sophonauth.com',
-  [sophonOSTestnet.id]: 'https://auth.staging.sophonauth.com',
+  [sophonOSTestnet.id]: 'http://localhost:4001',
 };
 
 /**
@@ -212,7 +210,7 @@ export const CHAIN_CONTRACTS: Record<ChainId, ContractAddresses> = {
   [sophonOSTestnet.id]: {
     session: zeroAddress,
     passkey: zeroAddress,
-    accountFactory: '0x0000006648ED9B2B842552BE63Af870bC74af837',
+    accountFactory: '0x5457Ce09A36cCd2b976497670979b90dC9465852',
     accountImplementation: '0x00000000383e8cBe298514674Ea60Ee1d1de50ac',
     bootstrap: '0x0000003eDf18913c01cBc482C978bBD3D6E8ffA3',
     accountPaymaster: zeroAddress,