+++ packages/account-react-native/src/token/me.ts
@@ -1,0 +1,21 @@
+import { AccountAuthAPIURL, type ChainId } from '@sophon-labs/account-core';
+import { getAccessToken } from './access-token';
+
+export const getMe = async (chainId: ChainId) => {
+  const token = await getAccessToken(chainId);
+  if (!token?.value) {
+    throw new Error('No access token found to use');
+  }
+
+  const response = await fetch(`${AccountAuthAPIURL[chainId]}/auth/me`, {
+    headers: {
+      authorization: `Bearer ${token.value}`,
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error(`Failed to get user info /me: ${response.statusText}`);
+  }
+
+  return await response.json();
+};